+++ src/main/java/jscover/instrument/BranchInstrumentor.java
@@ -351,8 +351,8 @@ import java.util.*;
 import static java.lang.String.format;
 
 public class BranchInstrumentor implements NodeVisitor {
-    private static final String initBranchLine = "  _$jscoverage['%s'].branchData[%d] = [];\n";
-    private static final String initBranchCondition = "  _$jscoverage['%s'].branchData[%d][%d] = new BranchData();\n";
+    private static final String initBranchLine = "  _$jscoverage['%s'].branchData['%d'] = [];\n";
+    private static final String initBranchCondition = "  _$jscoverage['%s'].branchData['%d'][%d] = new BranchData();\n";
 
     private static int functionId = 1;
     private BranchStatementBuilder branchStatementBuilder = new BranchStatementBuilder();
@@ -501,7 +501,7 @@ public class BranchInstrumentor implements NodeVisitor {
 
     protected String getJsLineInitialization() {
         StringBuilder sb = new StringBuilder(format("if (! _$jscoverage['%s'].branchData) {\n", uri));
-        sb.append(format("  _$jscoverage['%s'].branchData = [];\n", uri));
+        sb.append(format("  _$jscoverage['%s'].branchData = {};\n", uri));
         for (Integer line : lineConditionMap.keySet()) {
             sb.append(format(initBranchLine, uri, line));
             for (Integer condition: lineConditionMap.get(line))