+++ src/test-integration/resources/test-instrumented-file-system.js
@@ -95,28 +95,22 @@ function convertBranchDataConditionArrayToJSON(branchDataConditionArray) {
 
 function convertBranchDataLinesToJSON(branchData) {
     if (branchData === undefined) {
-        return '[]'
+        return '{}'
     }
-    var array = [];
-    var length = branchData.length;
-    for (var line = 0; line < length; line++) {
-        var branchDataObject = branchData[line];
-        if (branchDataObject === undefined || branchDataObject === null) {
-            value = 'null';
-        } else {
-            value = convertBranchDataConditionArrayToJSON(branchDataObject);
-        }
-        array.push(value);
+    var json = '';
+    for (var line in branchData) {
+        if (json !== '')
+            json += ','
+        json += '"' + line + '":' + convertBranchDataConditionArrayToJSON(branchData[line]);
     }
-    return '[' + array.join(',') + ']';
+    return '{' + json + '}';
 }
 
 function convertBranchDataLinesFromJSON(jsonObject) {
     if (jsonObject === undefined) {
-        return [];
+        return {};
     }
-    var length = jsonObject.length;
-    for (var line = 0; line < length; line++) {
+    for (var line in jsonObject) {
         var branchDataJSON = jsonObject[line];
         if (branchDataJSON !== null) {
             for (var conditionIndex = 0; conditionIndex < branchDataJSON.length; conditionIndex ++) {
@@ -130,37 +124,69 @@ function convertBranchDataLinesFromJSON(jsonObject) {
     return jsonObject;
 }
 function jscoverage_quote(s) {
-  return '"' + s.replace(/[\u0000-\u001f"\\\u007f-\uffff]/g, function (c) {
-    switch (c) {
-    case '\b':
-      return '\\b';
-    case '\f':
-      return '\\f';
-    case '\n':
-      return '\\n';
-    case '\r':
-      return '\\r';
-    case '\t':
-      return '\\t';
-    // IE doesn't support this
-    /*
-    case '\v':
-      return '\\v';
-    */
-    case '"':
-      return '\\"';
-    case '\\':
-      return '\\\\';
-    default:
-      return '\\u' + jscoverage_pad(c.charCodeAt(0).toString(16));
+    return '"' + s.replace(/[\u0000-\u001f"\\\u007f-\uffff]/g, function (c) {
+        switch (c) {
+            case '\b':
+                return '\\b';
+            case '\f':
+                return '\\f';
+            case '\n':
+                return '\\n';
+            case '\r':
+                return '\\r';
+            case '\t':
+                return '\\t';
+            // IE doesn't support this
+            /*
+             case '\v':
+             return '\\v';
+             */
+            case '"':
+                return '\\"';
+            case '\\':
+                return '\\\\';
+            default:
+                return '\\u' + jscoverage_pad(c.charCodeAt(0).toString(16));
+        }
+    }) + '"';
+}
+
+function getArrayJSON(coverage) {
+    var array = [];
+    if (coverage === undefined)
+        return array;
+
+    var length = coverage.length;
+    for (var line = 0; line < length; line++) {
+        var value = coverage[line];
+        if (value === undefined || value === null) {
+            value = 'null';
+        }
+        array.push(value);
     }
-  }) + '"';
+    return array;
+}
+
+function jscoverage_serializeCoverageToJSON() {
+    var json = [];
+    for (var file in _$jscoverage) {
+        var lineArray = getArrayJSON(_$jscoverage[file].lineData);
+        var fnArray = getArrayJSON(_$jscoverage[file].functionData);
+
+        json.push(jscoverage_quote(file) + ':{"lineData":[' + lineArray.join(',') + '],"functionData":[' + fnArray.join(',') + '],"branchData":' + convertBranchDataLinesToJSON(_$jscoverage[file].branchData) + '}');
+    }
+    return '{' + json.join(',') + '}';
+}
+
+
+function jscoverage_pad(s) {
+    return '0000'.substr(s.length) + s;
 }
 
 function jscoverage_html_escape(s) {
-    return s.replace(/[<>\&\"\']/g, function(c) {
-    return '&#' + c.charCodeAt(0) + ';';
-  });
+    return s.replace(/[<>\&\"\']/g, function (c) {
+        return '&#' + c.charCodeAt(0) + ';';
+    });
 }
 try {
   if (typeof top === 'object' && top !== null && typeof top.opener === 'object' && top.opener !== null) {
@@ -211,7 +237,7 @@ if (! _$jscoverage['test-simple.js'].functionData) {
   _$jscoverage['test-simple.js'].functionData = [];
 }
 if (! _$jscoverage['test-simple.js'].branchData) {
-  _$jscoverage['test-simple.js'].branchData = [];
+  _$jscoverage['test-simple.js'].branchData = {};
 }
 _$jscoverage['test-simple.js'].lineData[1]++;
 var x, y;